Move GridExpressionInput to new file
